Move sensible heat function into new fluxes.py and associated tests into new test_fluxes.py
