+++ src/CSET/operators/misc.py
@@ -615,76 +615,3 @@ def differentiate(
         return new_cubelist[0]
     else:
         return new_cubelist
-
-
-def sensible_heat_units(cubes, **kwargs):
-    """
-    Compute surface upward sensible heat flux from covariances.
-
-    EXPECTS (exactly one of each, preselected upstream):
-      - wt_covariance_2m
-      - air_temperature_rtd_1p2m
-      - barometric_pressure
-    UM cubes are passed through unchanged.
-    HEIGHT is treated as a *nominal* height (for UM selection / labelling),
-    not as a measurement height for Cardington.
-    """
-    from cf_units import Unit
-
-    Cp = 1004.67  # J kg-1 K-1
-    Rd = 287.05  # J kg-1 K-1
-    cubes = (
-        iris.cube.CubeList(cubes)
-        if not isinstance(cubes, iris.cube.CubeList)
-        else cubes
-    )
-
-    # --- Extract Cardington inputs explicitly listed upstream ---
-    if "CARDINGTON_VARNAMES" not in kwargs:
-        raise ValueError("sensible_heat_units requires CARDINGTON_VARNAMES")
-
-    wanted = set(kwargs["CARDINGTON_VARNAMES"].split(","))
-    selected = {c.var_name: c for c in cubes if c.var_name in wanted}
-    missing = wanted - set(selected)
-    if missing:
-        raise ValueError(
-            f"sensible_heat_units missing Cardington inputs: {sorted(missing)}"
-        )
-
-    wT = next(v for k, v in selected.items() if k.startswith("wt_covariance_"))
-    temp = next(v for k, v in selected.items() if k.startswith("air_temperature_rtd_"))
-    pressure = selected["pressure_barometric"]
-
-    # --- Unit handling ---
-    temp_K = temp.copy()
-    if temp_K.units is None or temp_K.units.is_unknown():
-        temp_K.units = Unit("degC")
-    temp_K.convert_units("K")
-
-    pres_Pa = pressure.copy()
-    if pres_Pa.units is None or pres_Pa.units.is_unknown():
-        pres_Pa.units = Unit("hPa")
-    pres_Pa.convert_units("Pa")
-
-    # --- Compute sensible heat flux ---
-    rho_air = pres_Pa.data / (Rd * temp_K.data)
-    shf = wT.copy()
-    shf.data = Cp * rho_air * wT.data
-    shf.units = "W m-2"
-    shf.rename("surface_upward_sensible_heat_flux_cardington")
-    shf.var_name = "surface_upward_sensible_heat_flux_cardington"
-
-    # --- Metadata: be explicit about mixed heights ---
-    shf.attributes["model_name"] = wT.attributes.get("model_name")
-    shf.attributes["cardington_measurement_heights"] = {
-        "wt_covariance": wT.var_name.split("_")[-1],
-        "air_temperature": temp.var_name.split("_")[-1],
-        "air_pressure": "1.2 m",
-    }
-    if "HEIGHT" in kwargs:
-        shf.attributes["nominal_height"] = f"{kwargs['HEIGHT']} m"
-
-    # --- Return: passthrough everything except Cardington inputs, plus SHF ---
-    out = iris.cube.CubeList(c for c in cubes if c.var_name not in wanted)
-    out.append(shf)
-    return out if len(out) > 1 else out[0]
